requirements.txt
asyncio
aiohttp
aiofiles
rich
discord-webhook[async]
feedparser
python-dotenv
python-dateutil